Refactor LessonTemplate into modular components and hooks
