+++ src/types/character.ts
@@ -1,0 +1,8 @@
+export interface Character {
+  id: string;
+  name: string;
+  avatar: string;
+  personality: string;
+  specialty: string;
+  color: string;
+}
